+++ MOP/src/MOP.cs
@@ -37,7 +37,7 @@ namespace MOP
         public override string Name => "MODERN OPTIMIZATION PLUGIN";
 #endif
         public override string Author => "Athlon"; //Your Username
-        public override string Version => "3.3.3"; //Version
+        public override string Version => "3.3.4"; //Version
         public const string SubVersion = ""; // NIGHTLY-yyyymmdd | BETA_x | RC_
         public override string UpdateLink => "https://github.com/Athlon007/MOP";
         public override byte[] Icon => Properties.Resources.icon;
@@ -203,6 +203,15 @@ namespace MOP
                 prompt.AddButton("OK", null);
                 prompt.AddButton("QUIT GAME", () => Application.Quit());
             }
+
+            if (MopSettings.AttemptedToFixTheGame && !MopSettings.AttemptedToFixTheGameRestart)
+            {
+                MopSettings.AttemptedToFixTheGameRestart = true;
+                BugReporter.Instance.RestartGame();
+                return;
+            }
+            MopSettings.AttemptedToFixTheGame = false;
+            MopSettings.AttemptedToFixTheGameRestart = false;
         }
 
         public override void ModSettingsLoaded()